pyproject.toml
[project]
name = "embeddable-ollama-chatbot-api"
version = "0.1.0"
description = "FastAPI proxy backend for Ollama with embeddable website widget integration"
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.0",
    "uvicorn[standard]>=0.40.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
manage-bots = "app.cli:main"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["app*"]

[dependency-groups]
dev = [
    "paramiko>=3.5.0",
    "pytest>=9.0.2",
]
